proper prompting
